+++ examples/web/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @angadie/example-web
 
+## 0.0.12
+
+### Patch Changes
+
+- @angadie/chittie@0.5.7
+
 ## 0.0.11
 
 ### Patch Changes